requirements.txt
python-dotenv==1.0.1
boto3==1.34.34
pandas==2.2.1
uuid6>=1.2.1

